backend/pyproject.toml
[project]
name = "biomon"
version = "0.1.0"
description = "Biosecurity Research Paper Monitor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "anthropic>=0.18.1",
    "arxiv>=2.1.0",
    "biopython>=1.83",
    "httpx>=0.26.0",
    "apscheduler>=3.10.4",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.26.0",
]

[project.scripts]
biomon = "app.scheduler:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

